+++ R/sdw.R
@@ -6,7 +6,7 @@
 #'
 #'
 #' @param verbose Logical, indicating whether to provide processing and retrieval messages. Defaults to FALSE
-#' @importFrom httr GET content accept_json http_status
+#' @import httr
 #' @importFrom purrr map_df
 #' @return returns a dataframe
 #' @export
@@ -27,7 +27,7 @@ echoSDWGetMeta <- function(verbose = FALSE){
   request <- httr::GET(getURL, httr::accept_json())
 
   ## Print status message, need to make this optional
-  if (verbose) {
+  if (isTRUE(verbose)) {
     message("Request URL:", getURL)
     message(httr::http_status(request))
   }
@@ -54,7 +54,9 @@ echoSDWGetMeta <- function(verbose = FALSE){
 #' Uses EPA's ECHO API: \url{https://echo.epa.gov/tools/web-services/facility-search-drinking-water#!/Safe_Drinking_Water/get_sdw_rest_services_get_systems}.
 #' @param verbose Logical, indicating whether to provide processing and retrieval messages. Defaults to FALSE
 #' @param ... Further arguments passed as query parameters in request sent to EPA ECHO's API. For more options see: \url{https://echo.epa.gov/tools/web-services/facility-search-drinking-water#!/Safe_Drinking_Water/get_sdw_rest_services_get_systems} for a complete list of parameter options. Examples provided below.
-#' @importFrom httr GET content accept_json http_status
+#' @importFrom purrr map
+#' @import httr
+#' @import dplyr
 #' @return returns a dataframe
 #' @export
 #'
@@ -81,6 +83,9 @@ echoSDWGetSystems <- function(verbose = FALSE, ...) {
     valuesList[["qcolumns"]] <- qcolumns
   }
 
+  # check if 1 and 2 are in, if not, insert and order
+  valuesList <- insertQColumns(valuesList)
+
   ## generate query the will be pasted into GET URL
   queryDots <- queryList(valuesList)
 
@@ -93,7 +98,7 @@ echoSDWGetSystems <- function(verbose = FALSE, ...) {
   request <- httr::GET(getURL, httr::accept_json())
 
   ## Print status message, need to make this optional
-  if (verbose) {
+  if (isTRUE(verbose)) {
     message("Request URL:", getURL)
     message(httr::http_status(request))
   }
@@ -105,9 +110,25 @@ echoSDWGetSystems <- function(verbose = FALSE, ...) {
   ## get qcolumns argument specific to this query
   qcolumns <- queryList(valuesList["qcolumns"])
 
+  ## Find out column types
+  colNums <- unlist(strsplit(valuesList[["qcolumns"]], split = ","))
+  colNums <- as.numeric(colNums)
+
+  ## ECHO always returns columns 1 and 2
+  ## regardless of the url request.
+  ## In order to correctly sort and identify column
+  ## types, insert 1 and 2 into the request so
+  ## metadat is looked up correctly
+  if (!1 %in% colNums) { colNums <- append(colNums, 1)}
+  if (!2 %in% colNums) { colNums <- append(colNums, 2)}
+  colNums <- sort(colNums)
+
+  colTypes <- columnsToParse(program = "sdw", colNums)
+
   ## call new function get_qid
   buildOutput <- getDownload(service = "sdw",
                              qid = qid,
-                             qcolumns = qcolumns)
+                             qcolumns = qcolumns,
+                             col_types = colTypes)
   return(buildOutput)
 }