+++ R/water.R
@@ -10,7 +10,8 @@
 #' @param verbose Logical, indicating whether to provide processing and retrieval messages. Defaults to FALSE
 #' @param \dots Further arguments passed as query parameters in request sent to EPA ECHO's API. For more options see: \url{https://echo.epa.gov/tools/web-services/facility-search-water#!/Facility_Information/get_cwa_rest_services_get_facility_info} for a complete list of parameter options. Examples provided below.
 #' @return returns a dataframe or simple features dataframe
-#' @importFrom httr GET content accept_json http_status
+#' @import httr
+#' @importFrom geojsonsf geojson_sf
 #'
 #' @export
 #' @examples \donttest{
@@ -52,61 +53,63 @@ echoWaterGetFacilityInfo <- function(output = "df",
       valuesList[["qcolumns"]] <- qcolumns
     }
 
+    # check if 1 and 2 are in, if not, insert and order
+    valuesList <- insertQColumns(valuesList)
+
     ## generate query the will be pasted into GET URL
     queryDots <- queryList(valuesList)
 
+    ## build the request URL statement
+    path <- "echo/cwa_rest_services.get_facility_info"
+    query <- paste("output=JSON", queryDots, sep = "&")
+    getURL <- requestURL(path = path, query = query)
+
+    ## Make the request
+    request <- httr::GET(getURL, httr::accept_json())
+
+    ## Print status message
+    if (isTRUE(verbose)) {
+      message("The formatted URL is: ", getURL)
+      message(httr::http_status(request))
+    }
+
+    info <- httr::content(request)
+
+    qid <- info[["Results"]][["QueryID"]]
+
+    ## build the output
+
+    ## get qcolumns argument specific to this query
+    qcolumns <- queryList(valuesList["qcolumns"])
+
+    ## Find out column types
+    colNums <- unlist(strsplit(valuesList[["qcolumns"]], split = ","))
+    colNums <- as.numeric(colNums)
+
+    colTypes <- columnsToParse(program = "cwa", colNums)
+
+    ## if df return output from air_rest_services.get_download
     if (output == "df") {
 
-        ## build the request URL statement
-        path <- "echo/cwa_rest_services.get_facility_info"
-        query <- paste("output=JSON", queryDots, sep = "&")
-        getURL <- requestURL(path = path, query = query)
-
-        ## Make the request
-        request <- httr::GET(getURL, httr::accept_json())
-
-        ## Print status message, need to make this optional
-        if (verbose) {
-            message("Request URL:", getURL)
-            message(httr::http_status(request))
-        }
-
-        info <- httr::content(request)
-
-        qid <- info[["Results"]][["QueryID"]]
-
-        ## build the output
-
-        ## get qcolumns argument specific to this query
-        qcolumns <- queryList(valuesList["qcolumns"])
-
-        buildOutput <- getDownload("cwa", qid, qcolumns)
+               buildOutput <- getDownload("cwa",
+                                   qid,
+                                   qcolumns,
+                                   col_types = colTypes)
         return(buildOutput)
 
         }
 
+    ## if df return output from air_rest_services.get_geojson
     if (output == "sf") {
 
-        ## build the request URL statement
-        path <- "echo/cwa_rest_services.get_facility_info"
-        query <- paste("output=GEOJSON", queryDots, sep = "&")
-        getURL <- requestURL(path = path, query = query)
+      buildOutput <- getGeoJson("cwa",
+                                qid,
+                                qcolumns)
+      ## Convert to sf dataframe
+      buildOutput <- geojsonsf::geojson_sf(buildOutput)
 
-        ## Make the request
-        request <- httr::GET(getURL, httr::accept_json())
+      return(buildOutput)
 
-        ## Print status message, need to make this optional
-        if (verbose) {
-            message("Request URL:", getURL)
-            message(httr::http_status(request))
-        }
-
-        ## Download GeoJSON as text
-        buildOutput <- httr::content(request, as = "text")
-
-        ## Convert to sf dataframe
-        buildOutput <- convertSF(buildOutput)
-        return(buildOutput)
     } else {
         stop("output argument = ", output,
              ", when it should be either 'df' or 'sf'")
@@ -122,7 +125,7 @@ echoWaterGetFacilityInfo <- function(output = "df",
 #' @param verbose Logical, indicating whether to provide processing and retrieval messages. Defaults to FALSE
 #'
 #' @return returns a dataframe
-#' @importFrom httr GET content accept_json http_status
+#' @import httr
 #' @importFrom purrr map_df
 #' @export
 #'
@@ -142,7 +145,7 @@ echoWaterGetMeta <- function(verbose = FALSE){
   request <- httr::GET(getURL, httr::accept_json())
 
   ## Print status message, need to make this optional
-  if (verbose) {
+  if (isTRUE(verbose)) {
     message("Request URL:", getURL)
     message(httr::http_status(request))
   }
@@ -167,7 +170,7 @@ echoWaterGetMeta <- function(verbose = FALSE){
 #' @param p_id Character string specify the identifier for the service. Required.
 #' @param verbose Logical, indicating whether to provide processing and retrieval messages. Defaults to FALSE
 #' @param ... Further arguments passed on as query parameters sent to EPA's ECHO API. For more options see: \url{https://echo.epa.gov/tools/web-services/effluent-charts#!/Effluent_Charts/get_eff_rest_services_get_effluent_chart}
-#' @importFrom httr GET content accept_json http_status
+#' @import httr
 #' @importFrom lubridate dmy
 #' @importFrom purrr map map_chr
 #' @importFrom tibble tibble
@@ -203,9 +206,9 @@ echoGetEffluent <- function(p_id, verbose = FALSE, ...) {
 
     request <- httr::GET(getURL, httr::accept_json())
 
-    if (verbose) {
-        message("Request URL:", getURL)
-        message(httr::http_status(request))
+    if (isTRUE(verbose)) {
+      message("Request URL:", getURL)
+      message(httr::http_status(request))
     }
 
     info <- httr::content(request)
@@ -288,7 +291,7 @@ echoGetEffluent <- function(p_id, verbose = FALSE, ...) {
 #' Search parameter codes for Clean Water Act permits on EPA ECHO
 #'
 #' Returns a dataframe of parameter codes and descriptions.
-#' @importFrom httr GET content accept_json http_status
+#' @import httr
 #' @importFrom purrr map_df
 #' @importFrom utils URLencode
 #'
@@ -338,7 +341,7 @@ echoWaterGetParams <- function(term = NULL, code = NULL, verbose = FALSE){
 
   request <- httr::GET(getURL, httr::accept_json())
 
-  if (verbose) {
+  if (isTRUE(verbose)) {
     message("Request URL:", getURL)
     message(httr::http_status(request))
   }